backend/requirements.txt
faiss-cpu==1.13.2
fastapi==0.135.3
numpy==2.4.4
PyMuPDF==1.27.2.2
python-dotenv==1.2.2
python-multipart==0.0.24
requests
scikit-learn==1.8.0
sentence-transformers==5.3.0
torch==2.11.0
transformers==5.5.0
uvicorn==0.44.0